+++ CHANGELOG.md
@@ -6,19 +6,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* `new` command
+* text record detail view
+
 ### Changed
 * reduced height of list items
 * configuration in a dict & optional values
+* organize by filename if no `Text_ID`
+* better feedback
 
 ## [0.1.1] - 2023-11-04
 
 ### Fixed
 * text loading, smoother interface overall
 
-## [0.1.0] - 2023-XX-XX
+## [0.1.0] - 2023-11-04
 
 Initial release
 
-[Unreleased]: https://github.com/fmatter/lingcorp/compare/v0.0.1...HEAD
-[0.1.1]: https://github.com/fmatter/lingcorp/compare/v0.0.1...v0.0.1
+[Unreleased]: https://github.com/fmatter/lingcorp/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/fmatter/lingcorp/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/fmatter/lingcorp/commit/insert_this_by_hand